+++ source/amd64/avx/parallel_amd64_avx.c
@@ -114,14 +114,14 @@ HIDDEN void *rbdParallelIdenticalWorkerAvx(void *arg)
     time = data->batchIdx * V4D;
 
     /* Align, if possible, to vector size */
-    if (((long)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
-        if (((long)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
+    if (((uintptr_t)&data->reliabilities[time] & (S1D * sizeof(double) - 1)) == 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (V2D * sizeof(double) - 1)) != 0) {
             /* Compute reliability of Parallel RBD at current time instant */
             rbdParallelIdenticalStepS1d(data, time);
             /* Increment current time instant */
             time += S1D;
         }
-        if (((long)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
+        if (((uintptr_t)&data->reliabilities[time] & (V4D * sizeof(double) - 1)) != 0) {
             /* Compute reliability of Parallel RBD at current time instant */
             rbdParallelIdenticalStepV2dSse2(data, time);
             /* Increment current time instant */
